Problem: Can't Comment On Problems
Solution:
Viewing comment threads. Some bit of refactor on posting threaded comments
